[PATCH] Kill warning in drivers/input/misc/uinput.c on IA64
From: Peter Chubb <peterc@gelato.unsw.edu.au>

Attached patch kills a warning when compiling on a 64-bit architecture
(ssize_t is long, not int)

+++ drivers/input/misc/uinput.c
@@ -226,7 +226,7 @@ exit:
 	return retval;
 }
 
-static int uinput_write(struct file *file, const char *buffer, size_t count, loff_t *ppos)
+static ssize_t uinput_write(struct file *file, const char *buffer, size_t count, loff_t *ppos)
 {
 	struct uinput_device	*udev = file->private_data;
